backend/requirements.txt
Django>=4.2
djangorestframework>=3.15.2
psycopg2-binary>=2.9.0
Pillow>=10.3.0
djangorestframework-simplejwt>=5.3.0
requests>=2.32.4
python-dotenv>=1.1.0
django-cors-headers>=4.7.0
django-imagekit>=5.0.0